traits support (not yet), check options vs validate
mark traits as a valid attribute option to test for, but not currently checked
(skipped).

Handle validating an attribute as a "thing" and its options at the same time
by interperting all keys of options to check that start with '-' as a key for
validate_thing().  This should allow the validate_*'s to pass off to
validate_attribute() without much violence.