validate_role ... -compose => 1 ... now tests anon class for role
validate_role($role, -compose => 1, ...) will now add a sanity test that the
composed subclass actually does $role when composing the arguments to pass to
validate_class().

Note this is a sanity check!  Of course the composed class will do $role, but
hey, it's *really* easy to lose track of one's sanity these days.